migrate PushInfo to styled-components
